+++ modules/mex/src/include/MxInteger.h
@@ -26,6 +26,7 @@
 #pragma once
 //=============================================================================
 #include "nlsMex_exports.h"
+#include "matrix.h"
 //=============================================================================
 #ifdef __cplusplus
 extern "C"
@@ -34,66 +35,37 @@ extern "C"
     //=============================================================================
     NLSMEX_IMPEXP
     bool
-    mxIsScalar(const mxArray* array_ptr);
+    mxIsInt8(const mxArray* pm);
     //=============================================================================
     NLSMEX_IMPEXP
     bool
-    mxIsEmpty(const mxArray* pm);
-    //=============================================================================
-    NLSMEX_IMPEXP
-    mwIndex
-    mxCalcSingleSubscript(const mxArray* pm, mwSize nsubs, mwIndex* subs);
-    //=============================================================================
-    NLSMEX_IMPEXP
-    int
-    mxSetDimensions(mxArray* pm, const mwSize* dims, mwSize ndim);
-    //=============================================================================
-    NLSMEX_IMPEXP
-    const mwSize*
-    mxGetDimensions(const mxArray* pm);
-    //=============================================================================
-    NLSMEX_IMPEXP
-    size_t
-    mxGetElementSize(const mxArray* pm);
-    //=============================================================================
-    NLSMEX_IMPEXP
-    mwSize
-    mxGetNumberOfDimensions(const mxArray* pm);
-    //=============================================================================
-    NLSMEX_IMPEXP
-    size_t
-    mxGetNumberOfElements(const mxArray* pm);
-    //=============================================================================
-    NLSMEX_IMPEXP
-    size_t
-    mxGetM(const mxArray* pm);
-    //=============================================================================
-    NLSMEX_IMPEXP
-    size_t
-    mxGetN(const mxArray* pm);
-    //=============================================================================
-    NLSMEX_IMPEXP
-    void
-    mxSetM(mxArray* pm, mwSize m);
-    //=============================================================================
-    NLSMEX_IMPEXP
-    void
-    mxSetN(mxArray* pm, mwSize n);
+    mxIsInt16(const mxArray* pm);
     //=============================================================================
     NLSMEX_IMPEXP
     bool
-    mxIsNumeric(const mxArray* pm);
+    mxIsInt32(const mxArray* pm);
     //=============================================================================
     NLSMEX_IMPEXP
     bool
-    mxIsComplex(const mxArray* pm);
+    mxIsInt64(const mxArray* pm);
     //=============================================================================
     NLSMEX_IMPEXP
     bool
-    mxIsFromGlobalWS(const mxArray* pm);
+    mxIsUint8(const mxArray* pm);
     //=============================================================================
-
+    NLSMEX_IMPEXP
+    bool
+    mxIsUint16(const mxArray* pm);
     //=============================================================================
+    NLSMEX_IMPEXP
+    bool
+    mxIsUint32(const mxArray* pm);
+    //=============================================================================
+    NLSMEX_IMPEXP
+    bool
+    mxIsUint64(const mxArray* pm);
+    //=============================================================================
+//=============================================================================
 #ifdef __cplusplus
 }
 #endif